requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
playwright==1.39.0
Pillow==10.0.1
reportlab==4.0.5